fixed all code
