+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Tue Jul 19 11:27:14 EAT 2022
-gradle.version=7.4
+#Tue Jul 19 12:45:06 EAT 2022
+gradle.version=7.2
